core is the proxy for frameworks
