+++ src/Component.js
@@ -1,49 +1,50 @@
 import inherit from 'inherit';
-import Bem from './Bem';
 
 let uniqCount = 0;
 const noop = () => {},
     prop = name => props => props[name];
 
-export default inherit(Bem, {
-    __constructor() {
-        this.__base(...arguments);
-        this.willInit(this.props);
-    },
+export default function(Bem) {
+    return inherit(Bem, {
+        __constructor() {
+            this.__base(...arguments);
+            this.willInit(this.props);
+        },
 
-    willInit : noop,
-    tag : noop,
-    attrs : noop,
-    style : noop,
-    mods : noop,
-    addMix : noop,
-    cls : prop('cls'),
-    mix : prop('mix'),
-    content : prop('children'),
+        willInit : noop,
+        tag : noop,
+        attrs : noop,
+        style : noop,
+        mods : noop,
+        addMix : noop,
+        cls : prop('cls'),
+        mix : prop('mix'),
+        content : prop('children'),
 
-    render() {
-        const { props } = this,
-            attrs = this.attrs(props) || {},
-            style = this.style(props) || {},
-            res = this.__render({
-                addBemClassName : this.addBemClassName,
-                tag : this.tag(props),
-                attrs : { ...attrs, style : { ...attrs.style, ...style } },
-                block : this.block,
-                elem : this.elem,
-                mods : this.mods(props),
-                mix : [this.mix(props), this.addMix(props)],
-                cls : this.cls(props),
-                children : this.content(props, props.children)
-            });
+        render() {
+            const { props } = this,
+                attrs = this.attrs(props) || {},
+                style = this.style(props) || {},
+                res = this.__render({
+                    addBemClassName : this.addBemClassName,
+                    tag : this.tag(props),
+                    attrs : { ...attrs, style : { ...attrs.style, ...style } },
+                    block : this.block,
+                    elem : this.elem,
+                    mods : this.mods(props),
+                    mix : [this.mix(props), this.addMix(props)],
+                    cls : this.cls(props),
+                    children : this.content(props, props.children)
+                });
 
-        return this.wrap? this.wrap(res) : res;
-    },
+            return this.wrap? this.wrap(res) : res;
+        },
 
-    generateId(key = 'uniq') {
-        this.__uniqId = this.__uniqId || {};
-        return this.__uniqId[key]
-            ? this.__uniqId[key]
-            : (this.__uniqId[key] = `${key}${++uniqCount}`);
-    }
-});
+        generateId(key = 'uniq') {
+            this.__uniqId = this.__uniqId || {};
+            return this.__uniqId[key]
+                ? this.__uniqId[key]
+                : (this.__uniqId[key] = `${key}${++uniqCount}`);
+        }
+    });
+}
