+++ src/presets/React.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+
+export default {
+    Render : React.createElement.bind(React),
+    Base : Component,
+    classAttribute : 'className',
+    PropTypes
+};
